pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rift-eval"
version = "0.1.0"
description = "Detect behavioral regressions between LLM model versions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Shah Baig" },
]
keywords = ["llm", "evaluation", "drift", "regression", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "rich>=13.7",
    "pydantic>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "pyright>=1.1",
]

[project.scripts]
rift = "rift.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rift"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
